refactor: CartItem, CartItemList 결합
